feat(phase-2): split kpi-report into legacy/v2 with feature-flag routing
